+++ src/Pages/folder/FolderPage.js
@@ -1,8 +1,7 @@
 import useAsync from "../../Hooks/useAsync";
 import { getUser } from "../../api";
 
-
-import Footer from "../../components/Footer/Footer";
+import Footer from "../../components/common/Footer";
 
 import Nav from "../../components/common/Nav";
