+++ src/Pages/shared/SharedPage.js
@@ -2,7 +2,7 @@ import useAsync from "../../Hooks/useAsync";
 import getSample from "../../api";
 import Folder from "../../components/Folder";
 
-import Footer from "../../components/Footer/Footer";
+import Footer from "../../components/common/Footer";
 import Nav from "../../components/common/Nav";
 
 function SharedPage() {
